+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## [Unreleased]
 
+## 1.2.0 (2019-04-04)
+
+* changed store object mutation to export functions  
+* added middleware functionality  
+* changed argument order of inject function
+* minor code improvements  
+
 ## 1.1.1 (2019-03-29)
 
 * removed some legacy code